+++ packages/fast-check/src/utils/iterator.ts
@@ -7,7 +7,6 @@ class Nil<T> implements IterableIterator<T> {
   }
 }
 
-/** @internal */
 export const nil: IteratorObject<any> = Iterator.from(new Nil<any>());
 
 export function getNthOrLast<T>(it: IterableIterator<T>, nth: number): T | null {
@@ -30,4 +29,21 @@ export function* joinAll<T, U, V>(its: IteratorObject<T, U, V>[]): IteratorObjec
     last = cur.value;
   }
   return last;
+}
+
+class LazyIterator<T> extends Iterator<T> {
+  private it?: IteratorObject<T>;
+  constructor(private readonly producer: () => IteratorObject<T>) {
+    super();
+  }
+  next(...[value]: [] | [any]): IteratorResult<T, any> {
+    if (this.it === undefined) {
+      this.it = this.producer();
+    }
+    return this.it.next(value);
+  }
+}
+
+export function makeLazy<T>(producer: () => IteratorObject<T>): IteratorObject<T> {
+  return new LazyIterator(producer);
 }
